Support update device and changed library version

+++ housetainer-adapter-persistence/src/main/kotlin/com/housetainer/adapter/persistence/repository/r2dbc/DeviceR2DBCRepository.kt
@@ -3,7 +3,7 @@ package com.housetainer.adapter.persistence.repository.r2dbc
 import com.housetainer.adapter.persistence.repository.entity.DeviceEntity
 import org.springframework.data.repository.kotlin.CoroutineCrudRepository
 
-interface DeviceR2DBCRepository : CoroutineCrudRepository<DeviceEntity, Int> {
+interface DeviceR2DBCRepository : CoroutineCrudRepository<DeviceEntity, Long> {
 
     suspend fun findByDeviceId(deviceId: String): DeviceEntity?
 }
